chore: adding dotenv
the initial `.gitignore` file references ".env*.local" files, which ran counter to the initial setup which included just a `.env` file. this was obviously to enable tracking to include the file for the assignment, but we don't want that anymore so we need to rename it. when doing that, though, in order to make sure the variables from the `.env.local` file are included in project files we need to use dotenv to tell it where to look for the files